feat(engine): CW/CH en float + repositionnement W/H/CW/CH sur graphics.canvas
CW/CH deviennent des FLOAT (W/2, H/2 exacts, ex. 400.5 pour largeur impaire).
graphics.canvas(w, h) recalcule desormais W/H/CW/CH sur les dimensions
logiques reelles du canvas (via setGlobal), au lieu de rester figes sur le
defaut window. Maj CLAUDE.md et tutoriel.

+++ src/vm.cpp
@@ -1541,10 +1541,10 @@ void VM::execute(Chunk chunk) {
                 globals[gi] = Value(win_h);
                 globals_init[gi] = true;
             } else if (owned_chunk.identifiers[gi] == "CW") {
-                globals[gi] = Value(win_w / 2);
+                globals[gi] = Value((double)win_w / 2.0);
                 globals_init[gi] = true;
             } else if (owned_chunk.identifiers[gi] == "CH") {
-                globals[gi] = Value(win_h / 2);
+                globals[gi] = Value((double)win_h / 2.0);
                 globals_init[gi] = true;
             }
         }
